Fixed table creation when global indexes contain non-indexed range keys.
When creating a table, dynamoose was only passing in key and
indexed attributes in the AttributeDefinitions parameter.
If a range key on a global index was not one of those attributes,
the table creation would fail because the range key was not
specified in the AttributeDefinitions list.

This fix looks for range keys on global secondary indexes
and appends them to the AttributeDefinitions list.