feat(ci/dev): move off grunt, make the ci config the same as local
This is focused on moving from two separate test runs, to a single
script to do the work. The value of these changes comes in consistency
in environments and ease of local dev.

Removing grunt for a few reasons:

  1. grunt as a runner for something simple is over kill
  1. as we are focused on node 8+ we can just use npm to mamange it
  1. one less thing to understand when working in the repo

These reasons, along with grunt's age and general support decline mean
to me that we can modernize and simplify. Hopefully that makes sense.