fix: change originalItem from being static
Original item was being held in `$__` which was located on the prototype
of the model. This meant that this value was being shared for all
instances of model.

To resolve this it is now stored on the model instance. To ensure that
it is not exposed to the user other than through `getOriginalItem` it is
stored as a Symbol.

This also required increased the allowed lines for the model test file.