Fix dynamic require of ModelStore
We encountered a strange bug which is hard to pin down, but
this particular commit has fixed it. Current hypothesis:

The instanceof check in ModelStore.ts is not operating correctly in
some environments. Suspect the inline require of ModelStore in Model.ts
is causing multiple instances of the ModelStore class to be loaded.

Passing the ModelStore singleton into the Model as a parameter
(and hence removing the dynamic require) appears to have solved this
issue.