Update import statment
I'm not sure if I'm correct here. Still, I kept getting the error  "Cannot read properties of undefined (reading 'ddb')" when running dynamoose.aws.ddb.local() to be consistent with the following page in the documentation I would only be able to run this method if I imported the default export.